+++ doc/Changelog.md
@@ -3,6 +3,8 @@ Changelog {#Changelog}
 
 ## Deflect 0.9 (git master)
 
+* [65](https://github.com/BlueBrain/Deflect/pull/65):
+  QmlStreamer: New 'streamname' command line option.
 * [64](https://github.com/BlueBrain/Deflect/pull/64):
   QmlStreamer: Fix event forwarding, implement wheel event support  
 * [60](https://github.com/BlueBrain/Deflect/pull/60):